% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{allegro_categories}
\alias{allegro_categories}
\title{Scrapping all allegro categories}
\usage{
allegro_categories()
}
\value{
Tibble of categories
}
\description{
Scrapping all allegro categories
}
\examples{
allegro_categories()

}
