% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{allegro_category_dump}
\alias{allegro_category_dump}
\title{Scrapping auctions by category}
\usage{
allegro_category_dump(category, file, order = "m", pageLimit = -1)
}
\arguments{
\item{file}{A file name}

\item{order}{One of values from allegro_order}

\item{pageLimit}{Page limit (-1 means all)}

\item{string}{A category}
}
\description{
Scrapping auctions by category
}
\examples{
allegro_category_dump("elektronika-67193", "elektronika-67193.xlsx", allegro_order$trafnosc, 1)

}
