% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{allegro_order}
\alias{allegro_order}
\title{Allegro sorting orders}
\format{An object of class \code{list} of length 8.}
\usage{
allegro_order
}
\description{
\describe{
  \item{trafnosc}{Sorting by accuracy}
  \item{cena_rosnaco}{Sorting by price ascending}
  \item{cena_malejaco}{Sorting by price descending}
  \item{cena_z_przesylka_rosnaco}{Sorting by price with supply ascending}
  \item{cena_z_przesylka_malejaco}{Sorting by price with supply descending}
  \item{popularnosc}{Sorting by popularity}
  \item{czas_do_konca}{Sorting by time to the end ascending}
  \item{czas_dodania}{Sorting by time of auction}
}
}
\examples{
allegro_order$trafnosc

}
\keyword{datasets}
