% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{allegro_page_dump}
\alias{allegro_page_dump}
\title{Dump request to prettified json file}
\usage{
allegro_page_dump(base_url, query, file)
}
\arguments{
\item{base_url}{A url}

\item{query}{Query list}

\item{file}{A file name}
}
\description{
Dump request to prettified json file
}
\examples{
allegro_page_dump("https://allegro.pl/mapa-strony/kategorie", list(), "kategorie.json")

}
