% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{allegro_serch_dump}
\alias{allegro_serch_dump}
\title{Scrapping auctions by search value}
\usage{
allegro_serch_dump(string, file, order = "m", pageLimit = -1)
}
\arguments{
\item{string}{A search value}

\item{file}{A file name}

\item{order}{One of values from \link{allegro_order}}

\item{pageLimit}{Page limit (-1 means all)}
}
\description{
Scrapping auctions by search value
}
\examples{
allegro_serch_dump("herbata", "herbata.xlsx", allegro_order$trafnosc, 1)

}
