% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{allegro_user_dump}
\alias{allegro_user_dump}
\title{Scrapping auctions by user auctions}
\usage{
allegro_user_dump(user, file, order = "m", pageLimit = -1)
}
\arguments{
\item{file}{A file name}

\item{order}{One of values from allegro_order}

\item{pageLimit}{Page limit (-1 means all)}

\item{string}{A user name}
}
\description{
Scrapping auctions by user auctions
}
\examples{
allegro_user_dump("X-KOM_PL", "X-KOM_PL.xlsx", allegro_order$trafnosc, 1)

}
