% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{rallegroapi}
\alias{rallegroapi}
\title{A package for scrapping data from allegro.pl.}
\description{
A package for scrapping data from allegro.pl.
}
\section{Functions}{

\code{\link{allegro_page_dump}}

\code{\link{allegro_serch_dump}}

\code{\link{allegro_category_dump}}

\code{\link{allegro_user_dump}}

\code{\link{allegro_categories}}
}

